finish ex. 2.5 |Course component in seperate module
